blog_django_graphql_react_relay/film_api/film_database/films/serializers.py
from rest_framework import serializers

from .models import Film


class FilmSerializer(serializers.ModelSerializer):
    class Meta:
        model = Film
        fields = ('title', 'air_date', 'actors', 'rating', 'id')
